+++ page/src/components/Footer.tsx
@@ -1,0 +1,12 @@
+const Footer = () => {
+  return (
+    <footer className="site-footer">
+      <span>Developed by </span>
+      <a href="https://github.com/JooLuiz" target="_blank" rel="noreferrer">
+        @JooLuiz
+      </a>
+    </footer>
+  );
+};
+
+export default Footer;